% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_roster}
\alias{get_roster}
\title{Get roster for a team}
\usage{
get_roster(team_id, league, sport = NULL)
}
\arguments{
\item{team_id}{Character, ESPN team ID}

\item{league}{Character, league abbreviation (nfl, nba, nhl, mlb)}

\item{sport}{Character, sport name (optional, will be inferred from league)}
}
\value{
data.table with roster information
}
\description{
Get roster for a team
}
\examples{
\dontrun{
# Get Lakers roster
lakers_roster <- get_roster("1610612747", "nba")

# Get Patriots roster
patriots_roster <- get_roster("17", "nfl")
}
}
