% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_teams}
\alias{get_teams}
\title{Get teams for a league}
\usage{
get_teams(league, sport = NULL)
}
\arguments{
\item{league}{Character, league abbreviation (nfl, nba, nhl, mlb)}

\item{sport}{Character, sport name (optional, will be inferred from league)}
}
\value{
data.table with team information
}
\description{
Get teams for a league
}
\examples{
\dontrun{
# Get all NBA teams
nba_teams <- get_teams("nba")

# Get all NFL teams
nfl_teams <- get_teams("nfl")
}
}
